requirements.txt
flask
pillow
scikit-learn
gunicorn